pyproject.toml
[tool.poetry]
name = "toolkit-tester2"
version = "0.1.0"
description = ""
authors = ["Ben Petree <ben.petree@cognite.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
cognite-toolkit = "^0.2.6"
cognite-extractor-utils = "^7.2.2"
cognite-sdk = "^7.51.1"
orjson = "^3.10.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
